Update cells CSV file
