+++ app/prompts.py
@@ -1,0 +1,6 @@
+
+
+
+
+
+STUDENT_QUERY = "What is the student's name? What is their GW ID?"
